Update index.html
since you added 

```
<span class=a href="#"><i class="fab fa-twitter"></i></a></li>"
```

you didn't seem to import font awesome icons so this won't work and the class is not in quotations it should be...the html here is a mess. Did you try this out before pushing?

also h2 is big do h6 if ever

and when naming classes and id please make it more declarative not `f4` etc.